pyproject.toml
[project]
name = "post"
version = "0.1.0"
description = "PoST"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "causal-conv1d>=1.6.0",
    "datasets>=4.5.0",
    "einops>=0.8.2",
    "lm-eval>=0.4.11",
    "mamba-ssm>=2.3.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
    "wandb>=0.24.1",
    "flash-linear-attention>=0.4.1",
    "wonderwords>=3.0.1",
    "nltk>=3.9.2",
]

[tool.setuptools.packages.find]
include = ["models*"]

[tool.uv.sources]
flash-linear-attention = { git = "https://github.com/fla-org/flash-linear-attention.git", branch = "main" }
